requirements.txt
langchain
langchain-core
langchain-openai
langchain-pinecone
openai
pinecone-client
pypdf2
python-dotenv
streamlit
streamlit-chat